app/src/main/java/com/example/calco/logic/persistent/converters/DateTimeConverter.java
package com.example.calco.logic.persistent.converters;

import java.time.Instant;
import java.time.LocalDateTime;
import java.time.OffsetDateTime;
import java.time.ZoneId;

public class DateTimeConverter {
    public static LocalDateTime timeFromUtcMillis(Long millis) {
        return LocalDateTime.ofInstant(Instant.ofEpochMilli(millis), ZoneId.systemDefault());
    }

    public static Long timeToUtcMillis(LocalDateTime dateTime) {
        return dateTime.toInstant(OffsetDateTime.now().getOffset()).toEpochMilli();
    }
}
